requirements.txt
django-rest-framework==0.1.0 
djangorestframework==3.6.4
django-cors-headers==2.1.0
numpy==1.13.1
pillow==4.2.1
keras==1.0.0
tensorflow==1.3.0
matplotlib==2.0.2
requests==2.18.4